fix(radiant): import SSR runtime from consolidated entrypoint
Merge from legacy-hardening restored radiant-element-ssr-runtime imports
after this branch removed that shim in favor of radiant-element-ssr.ts.
